Implement minimization / maximization, optimize chart algorithms, update NuGet packages
