+++ src/FrontEASE.Client/FrontEASE.Client.csproj
@@ -23,20 +23,20 @@
 	</ItemGroup>
 
   <ItemGroup>
-    <PackageReference Include="AutoMapper" Version="16.0.0" />
+    <PackageReference Include="AutoMapper" Version="16.1.0" />
     <PackageReference Include="Blazor-ApexCharts" Version="6.1.0" />
     <PackageReference Include="Blazored.LocalStorage" Version="4.5.0" />
     <PackageReference Include="Blazored.Toast" Version="4.2.1" />
-    <PackageReference Include="Blazorise.Analyzers" Version="2.0.1">
+    <PackageReference Include="Blazorise.Analyzers" Version="2.0.2">
       <PrivateAssets>all</PrivateAssets>
       <IncludeAssets>runtime; build; native; contentfiles; analyzers; buildtransitive</IncludeAssets>
     </PackageReference>
-    <PackageReference Include="Blazorise.Bootstrap5" Version="2.0.1" />
-    <PackageReference Include="Blazorise.Components" Version="2.0.1" />
-    <PackageReference Include="Blazorise.DataGrid" Version="2.0.1" />
-    <PackageReference Include="Blazorise.Icons.FontAwesome" Version="2.0.1" />
-    <PackageReference Include="Blazorise.Markdown" Version="2.0.1" />
-    <PackageReference Include="Markdig" Version="1.1.0" />
+    <PackageReference Include="Blazorise.Bootstrap5" Version="2.0.2" />
+    <PackageReference Include="Blazorise.Components" Version="2.0.2" />
+    <PackageReference Include="Blazorise.DataGrid" Version="2.0.2" />
+    <PackageReference Include="Blazorise.Icons.FontAwesome" Version="2.0.2" />
+    <PackageReference Include="Blazorise.Markdown" Version="2.0.2" />
+    <PackageReference Include="Markdig" Version="1.1.1" />
     <PackageReference Include="Microsoft.AspNetCore.Components.Authorization" Version="10.0.3" />
     <PackageReference Include="Microsoft.AspNetCore.Components.WebAssembly" Version="10.0.3" />
     <PackageReference Include="Microsoft.AspNetCore.Components.WebAssembly.DevServer" Version="10.0.3" PrivateAssets="all" />
