+++ src/FrontEASE.Client/_Imports.razor
@@ -77,6 +77,7 @@
 @using FrontEASE.Client.Services.HelperServices.UI.Manage
 @using FrontEASE.Client.Services.HelperServices.UI.Operations
 @using FrontEASE.Client.Services.HelperServices.UI.Signals
+@using FrontEASE.Client.Services.HelperServices.UI.Charts
 @using FrontEASE.Client.Shared.Components.Forms
 @using FrontEASE.Client.Shared.Components.Modals
 @using FrontEASE.Client.Shared.Components.Utility
