+++ src/FrontEASE.Shared/Data/Enums/Tasks/Config/OptimizationGoalType.cs
@@ -1,0 +1,18 @@
+﻿namespace FrontEASE.Shared.Data.Enums.Tasks.Config
+{
+    /// <summary>
+    /// Enum representing the optimization goal
+    /// </summary>
+    public enum OptimizationGoalType
+    {
+        /// <summary>
+        /// Minimization of the metric (lower is better).
+        /// </summary>
+        MINIMIZATION,
+
+        /// <summary>
+        /// Maximization of the metric (higher is better).
+        /// </summary>
+        MAXIMIZATION
+    }
+}
